More refactoring to accomodate encoding

+++ include/aws/mqtt/private/v5/mqtt5_utils.h
@@ -12,13 +12,13 @@
 
 struct aws_byte_buf;
 
-#define AWS_MQTT5_MAXIMUM_VARIABLE_LENGTH_INTEGER ((uint32_t)268435455)
+#define AWS_MQTT5_MAXIMUM_VARIABLE_LENGTH_INTEGER 268435455
 
 AWS_EXTERN_C_BEGIN
 
 int aws_mqtt5_encode_variable_length_integer(struct aws_byte_buf *buf, uint32_t value);
 
-int aws_mqtt5_get_variable_length_encode_size(uint32_t value, uint32_t *encode_size);
+int aws_mqtt5_get_variable_length_encode_size(size_t value, size_t *encode_size);
 
 /**
  * Converts a disconnect reason code into the Reason Code Name, as it appears in the mqtt5 spec.
